# shared/openapi.yaml
openapi: 3.1.0
info:
  title: Artifact Atlas API
  version: 0.2.0
  description: >
    Local-first asset graph, Artifact BOM, and context-pack builder API.
    All API/storage enum values are canonical and stable; UI labels may differ.
    Sensitive assets default to metadata-only agent access.
    Canonical promotion is human-gated unless policy explicitly permits automation.
  contact:
    name: Artifact Atlas Project
  license:
    name: Proprietary
servers:
  - url: http://localhost:8000
    description: Local development server
...
paths:
...
  /api/templates:
    get:
      operationId: listTemplates
      tags: [templates]
      summary: List artifact templates
      parameters:
        - $ref: "#/components/parameters/cursor"
        - $ref: "#/components/parameters/limit"
        - name: status
          in: query
          schema:
            $ref: "#/components/schemas/TemplateStatus"
        - name: template_type
          in: query
          schema:
            $ref: "#/components/schemas/TemplateType"
      responses:
        "200":
          description: Paginated template list
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/CursorPage"
                  - type: object
                    properties:
                      items:
                        type: array
                        items:
                          $ref: "#/components/schemas/Template"
    post:
      operationId: createTemplate
      tags: [templates]
      summary: Create a new artifact template
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TemplateCreate"
      responses:
        "201":
          description: Template created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Template"
        "422":
          $ref: "#/components/responses/ValidationError"
  /api/templates/{templateId}:
    parameters:
      - $ref: "#/components/parameters/templateId"
    get:
      operationId: getTemplate
      tags: [templates]
      summary: Get a single template with domains and slots
      responses:
        "200":
          description: Template detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TemplateDetail"
        "404":
          $ref: "#/components/responses/NotFound"
    patch:
      operationId: updateTemplate
      tags: [templates]
      summary: Partially update a template
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TemplateUpdate"
      responses:
        "200":
          description: Updated template
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Template"
        "404":
          $ref: "#/components/responses/NotFound"
        "422":
          $ref: "#/components/responses/ValidationError"
  /api/templates/{templateId}/duplicate:
    parameters:
      - $ref: "#/components/parameters/templateId"
    post:
      operationId: duplicateTemplate
      tags: [templates]
      summary: Duplicate a template
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                  description: Name for the duplicate; defaults to "Copy of <original name>"
      responses:
        "201":
          description: Duplicate template created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Template"
        "404":
          $ref: "#/components/responses/NotFound"
  /api/templates/{templateId}/preview:
    parameters:
      - $ref: "#/components/parameters/templateId"
    get:
      operationId: previewTemplate
      tags: [templates]
      summary: Preview the slot structure of a template
      description: Returns domains and slots without applying to any project.
      responses:
        "200":
          description: Template slot preview
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TemplatePreview"
        "404":
          $ref: "#/components/responses/NotFound"
  /api/templates/{templateId}/impact:
    parameters:
      - $ref: "#/components/parameters/templateId"
    get:
      operationId: getTemplateImpact
      tags: [templates]
      summary: Preview impact of applying a template to a project
      description: >
        Returns slot counts for what would be added (new_slots) vs. conflicting with
        existing BOM slots (conflict_slots) before the user confirms application.
        Used by the Apply Template Wizard to show a merge preview.
      parameters:
        - name: project_id
          in: query
          required: true
          schema:
            type: string
          description: Project to compute impact against
      responses:
        "200":
          description: Template impact preview
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TemplateImpact"
        "404":
          $ref: "#/components/responses/NotFound"
  /api/templates/{templateId}/publish:
    parameters:
      - $ref: "#/components/parameters/templateId"
    post:
      operationId: publishTemplate
      tags: [templates]
      summary: Explicitly publish a draft template
      description: >
        Transitions a template from experimental (draft) status to recommended (published).
        This is the required explicit gate before a custom template can be used in
        production BOMs. Idempotent if template is already published.
      responses:
        "200":
          description: Published template
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Template"
        "404":
          $ref: "#/components/responses/NotFound"

  # ---------------------------------------------------------------------------
  # BOM
  # ---------------------------------------------------------------------------
...
components:
...
  parameters:
...
    templateId:
      name: templateId
      in: path
      required: true
      description: Unique template identifier
      schema:
        type: string
        example: tmpl_new_product_app_v1
...
    cursor:
      name: cursor
      in: query
      description: Opaque cursor for next page (from previous response next_cursor)
      schema:
        type: string
    limit:
      name: limit
      in: query
      description: Maximum items per page
      schema:
        type: integer
        default: 50
        minimum: 1
        maximum: 200

  # ---------------------------------------------------------------------------
  # Responses
  # ---------------------------------------------------------------------------
  responses:
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
          example:
            error:
              code: not_found
              message: "The requested resource was not found."
              request_id: "req_abc123"
...
    ValidationError:
      description: Request validation failed
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ValidationErrorEnvelope"
...
  schemas:
...
    TemplateStatus:
      type: string
      enum: [core, recommended, optional, experimental, deprecated]
      description: Canonical artifact template lifecycle status
    TemplateType:
      type: string
      enum: [product, architecture, gtm, research, design_system, client_proposal, platform_capability, custom]
...
    SlotPhase:
      type: string
      enum: [discovery, design, build, launch, operate, review, custom]
      nullable: true
...
    CursorPage:
      type: object
      required: [items, has_more]
      description: Cursor-paginated response wrapper
      properties:
        items:
          type: array
          items: {}
          description: Page of results (typed by the enclosing allOf)
        has_more:
          type: boolean
          description: Whether more pages exist after this one
        next_cursor:
          type: string
          nullable: true
          description: Opaque cursor to pass as ?cursor= for the next page
        total:
          type: integer
          nullable: true
          description: Total count when available (may be null for large collections)

    # ---- Error envelope ----
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code:
              type: string
              description: Machine-readable error code
              examples:
                - not_found
                - policy_denied
                - conflict
                - internal_error
            message:
              type: string
              description: Human-readable error description
            detail:
              type: object
              description: Additional structured context (varies by error type)
              nullable: true
            request_id:
              type: string
              nullable: true
    ValidationErrorEnvelope:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message, validation_errors]
          properties:
            code:
              type: string
              enum: [validation_error]
            message:
              type: string
            request_id:
              type: string
              nullable: true
            validation_errors:
              type: array
              items:
                type: object
                required: [field, message]
                properties:
                  field:
                    type: string
                    description: Dot-separated field path
                  message:
                    type: string
                  value:
                    description: The invalid value (if safe to include)

    # ---- Projects ----
...
    Template:
      type: object
      required: [id, name, slug, template_type, status, version, created_at, updated_at]
      properties:
        id:
          type: string
          example: tmpl_new_product_app_v1
        workspace_id:
          type: string
          nullable: true
        name:
          type: string
          example: New Product / App
        slug:
          type: string
          example: new-product-app
        description:
          type: string
          nullable: true
        template_type:
          $ref: "#/components/schemas/TemplateType"
        status:
          $ref: "#/components/schemas/TemplateStatus"
        version:
          type: string
          example: "1.0.0"
        created_by:
          type: string
          nullable: true
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
        metadata:
          type: object
          nullable: true
    TemplateDetail:
      allOf:
        - $ref: "#/components/schemas/Template"
        - type: object
          properties:
            domains:
              type: array
              items:
                $ref: "#/components/schemas/TemplateDomain"
    TemplateDomain:
      type: object
      required: [id, template_id, name, slug, display_order]
      properties:
        id:
          type: string
        template_id:
          type: string
        name:
          type: string
          example: Architecture
        slug:
          type: string
          example: architecture
        description:
          type: string
          nullable: true
        display_order:
          type: integer
        slots:
          type: array
          items:
            $ref: "#/components/schemas/TemplateSlot"
    TemplateSlot:
      type: object
      required: [id, template_id, domain_id, artifact_type_id, required, min_assets, display_order]
      properties:
        id:
          type: string
        template_id:
          type: string
        domain_id:
          type: string
        artifact_type_id:
          type: string
        phase:
          $ref: "#/components/schemas/SlotPhase"
        required:
          type: boolean
        min_assets:
          type: integer
          minimum: 0
          default: 1
        max_assets:
          type: integer
          nullable: true
        staleness_days:
          type: integer
          nullable: true
        linked_intenttree_node_pattern:
          type: string
          nullable: true
        display_order:
          type: integer
        rule_config:
          type: object
          nullable: true
    TemplateCreate:
      type: object
      required: [name, slug, template_type]
      properties:
        name:
          type: string
        slug:
          type: string
          pattern: "^[a-z0-9-]+$"
        description:
          type: string
          nullable: true
        template_type:
          $ref: "#/components/schemas/TemplateType"
        status:
          $ref: "#/components/schemas/TemplateStatus"
          default: experimental
        version:
          type: string
          default: "1.0.0"
        metadata:
          type: object
          nullable: true
    TemplateUpdate:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
          nullable: true
        status:
          $ref: "#/components/schemas/TemplateStatus"
        version:
          type: string
        metadata:
          type: object
          nullable: true
    TemplatePreview:
      type: object
      required: [template_id, domains]
      properties:
        template_id:
          type: string
        domains:
          type: array
          items:
            $ref: "#/components/schemas/TemplateDomain"
        total_slots:
          type: integer
        required_slots:
          type: integer
    TemplateImpact:
      type: object
      required: [template_id, template_name, total_slots, required_slots, optional_slots, new_slots, conflict_slots]
      description: >
        Impact preview returned by GET /api/templates/{templateId}/impact.
        Shows which slots would be net-new vs. conflicting with existing BOM slots.
      properties:
        template_id:
          type: string
        template_name:
          type: string
        total_slots:
          type: integer
          description: Total slot count in the template
        required_slots:
          type: integer
          description: Required slot count
        optional_slots:
          type: integer
          description: Optional slot count
        new_slots:
          type: integer
          description: Slots that would be created (no existing slot with same domain+type)
        conflict_slots:
          type: integer
          description: Slots that already exist in the project BOM (would be skipped on apply)
        conflict_details:
          type: array
          description: Per-slot conflict detail for UI merge preview
          items:
            type: object
            required: [domain, artifact_type_id, required]
            properties:
              domain:
                type: string
              artifact_type_id:
                type: string
              required:
                type: boolean

    # ---- BOM ----
...